Renamed lint.rb to linter.rb
