+++ lib/linter.rb
